Updated by moving loadbalancer to separate file.
