+++ index.js
@@ -1,60 +1,28 @@
 import express from 'express';
 import cors from 'cors';
 import helmet from 'helmet';
-import fetch from 'node-fetch';
 import shipRoutes from './routes/shipRoutes.js';
 import squadronRoutes from './routes/squadronRoutes.js';
 import upgradeRoutes from './routes/upgradeRoutes.js';
 import { errorHandler } from './middleware/errorHandler.js';
 import baseRoutes from './routes/baseRoutes.js';
 import objectiveRoutes from './routes/objectiveRoutes.js';
+import loadBalancer from './middleware/loadBalancer.js';
 
 const app = express();
 const PORT = process.env.PORT || 4000;
-
-const MAIN_API = 'https://api.swarmada.wiki';
-const BACKUP_API = 'https://api-backup.swarmada.wiki';
 
 app.use(cors());
 app.use(helmet());
 app.use(express.json());
 
-// Load balancer middleware
-const loadBalancer = async (req, res, next) => {
-  try {
-    const mainApiResponse = await fetch(`${MAIN_API}${req.url}`, {
-      method: req.method,
-      headers: req.headers,
-      body: req.method !== 'GET' && req.method !== 'HEAD' ? JSON.stringify(req.body) : undefined
-    });
-
-    if (mainApiResponse.ok) {
-      const data = await mainApiResponse.json();
-      return res.json(data);
-    }
-
-    // If main API fails, try backup
-    const backupApiResponse = await fetch(`${BACKUP_API}${req.url}`, {
-      method: req.method,
-      headers: req.headers,
-      body: req.method !== 'GET' && req.method !== 'HEAD' ? JSON.stringify(req.body) : undefined
-    });
-
-    if (backupApiResponse.ok) {
-      const data = await backupApiResponse.json();
-      return res.json(data);
-    }
-
-    // If both fail, pass to the next middleware (which will be your routes)
-    next();
-  } catch (error) {
-    console.error('Load balancer error:', error);
-    next();
-  }
-};
-
 // Apply load balancer to all routes
 app.use(loadBalancer);
+
+// Add this new route before your other route definitions
+app.get('/health', (req, res) => {
+  res.status(200).json({ status: 'healthy' });
+});
 
 // Your existing routes
 app.use('/', baseRoutes);
@@ -65,8 +33,17 @@ app.use('/api/objectives', objectiveRoutes);
 
 app.use(errorHandler);
 
+app.use((err, req, res, next) => {
+  console.error(err.stack);
+  res.status(500).send('Something broke!');
+});
+
 app.listen(PORT, () => {
   console.log(`Server running on port ${PORT}`);
+  checkApiHealth(MAIN_API).then(healthy => {
+    useMainApi = healthy;
+    console.log(`Initial health check: Using ${useMainApi ? 'Main' : 'Backup'} API.`);
+  });
 });
 
 export default app;
